Introduction to Criteria API
Gets records by specifying our criteria using the criteria API
Criteria it's like a "where" clause - we are specififying what is the criteria that you need to be fulfilled.
Criteria needs to know for which table or which entity you're imposing the restrction
If you skip criteria.add - you will get all the records.
criteria.add(Restrictions.eq("userName", "User 10")); - created a restriction,

#28